backend/fithub/requirements.txt
Django==5.2
djangorestframework==3.15.2
mysqlclient==2.1.0         # For MySQL database connection (sometimes fails)
PyMySQL==1.1.1             # For MySQL connection if mysqlclient won't work
psycopg2-binary>=2.9
python-dotenv==1.0.1       # To read .env file
drf-yasg>=1.21.7           # For swagger documentation
PyJWT==2.9.0               # JWT Token Authentication
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0 # For CORS headers

